trame/requirements.txt
fastapi
uvicorn
requests 
pyjwt 
python-dotenv 
psutil
virtualenv